+++ src/main/java/com/github/shCHO9801/climbing_record_app/user/dto/AuthRequest.java
@@ -1,4 +1,4 @@
-package com.github.shCHO9801.climbing_record_app.user.auth.dto;
+package com.github.shCHO9801.climbing_record_app.user.dto;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -7,8 +7,7 @@ import lombok.Setter;
 @Getter
 @Setter
 @Builder
-public class RegisterRequest {
+public class AuthRequest {
   private String username;
   private String password;
-  private String email;
 }
