+++ src/main/java/com/github/shCHO9801/climbing_record_app/user/dto/RegisterRequest.java
@@ -1,4 +1,4 @@
-package com.github.shCHO9801.climbing_record_app.user.auth.dto;
+package com.github.shCHO9801.climbing_record_app.user.dto;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -7,7 +7,8 @@ import lombok.Setter;
 @Getter
 @Setter
 @Builder
-public class AuthRequest {
+public class RegisterRequest {
   private String username;
   private String password;
+  private String email;
 }
