+++ src/test/java/com/github/shCHO9801/climbing_record_app/user/auth/controller/AuthControllerTest.java
@@ -8,8 +8,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.github.shCHO9801.climbing_record_app.user.auth.dto.AuthRequest;
-import com.github.shCHO9801.climbing_record_app.user.auth.dto.RegisterRequest;
+import com.github.shCHO9801.climbing_record_app.user.dto.AuthRequest;
+import com.github.shCHO9801.climbing_record_app.user.dto.RegisterRequest;
 import com.github.shCHO9801.climbing_record_app.user.entity.User;
 import com.github.shCHO9801.climbing_record_app.user.repository.UserRepository;
 import org.junit.jupiter.api.BeforeEach;
